Read-first variant with registered outputs hits ~ 227 MHz = manual instantiation
Instantiating the DP16KD blocks by hand reaches 223 MHz but there's so
much P&R variance that the difference might just be noise.

I'm very unconfortable with the fact that the DP16KD explicitly
disallow accessing the same location from both ports while the Verilog
version's semantics is perfectly defined in that case.  So how does
the Yosys deal with this?